feat: 홈 화면 경기 카드 섹션 구현 (#409)
